Split Changes from History
These two use practically the same template but there are small
changes that are more important now that we support the diff for
arbitrary revisions. The UI is now a big form starting with two
radiobuttons. Getting this and the list for all the other variants of
changes into one template was making things confusing so it was split
into two.